+++ analytics_toolkit/__init__.py
@@ -1,3 +1,3 @@
 """analytics_toolkit package."""
 
-__all__ = ["ab_utils", "dates", "excel", "general", "sql"]
+__all__ = ["ab_utils", "dates", "excel", "general", "sql", "sql_format"]